加了点文档，修了一个编译递归函数时候的bug

+++ print.c
@@ -1,3 +1,3 @@
 #include <stdio.h>
 int print_int(int64_t a) { return printf("%lld\n", a); }
-int print_float(double a) { return printf("%lf\n", a); }
+int print_float(float a) { return printf("%f\n", a); }
